+++ src/main/java/com/dekk/user/application/command/UserProfileUpdateCommand.java
@@ -2,5 +2,5 @@ package com.dekk.user.application.command;
 
 import com.dekk.user.domain.model.enums.Gender;
 
-public record UserProfileUpdateCommand(String nickname, Double height, Double weight, Gender gender) {
+public record UserProfileUpdateCommand(String nickname, Integer height, Integer weight, Gender gender) {
 }
